Use uncasted NULL pointers
C89 6.2.2.3 Pointers

An integral constant expression with the value 0, or such an expression
cast to type void *, is called a null pointer constant. If a null
pointer constant is assigned to or compared for equality to a pointer,
the constant is converted to a pointer of that type. Such a pointer,
called a null pointer, is guaranteedto compare unequal to a pointer to
any object or function.

Two null pointers, converted through possibly different sequences of
casts to pointer types, shall compare equal.